feature: Add sorting by number of ratings on 'Top Rated' page
Implemented a new feature that allows users to sort games/items by the number of ratings. This sorting functionality is available only on the "Top Rated" page and can be used to organize items by most or least number of ratings.

To maintain the style and usability of the site, the custom buttons for this new sorting feature were moved to a new row right below the original row of sorting buttons.

Note: Incrementing minor version to 0.3.0 to indicate the addition of new functionality.

Personal Note: Implementing the sorting feature took only a few minutes, thanks to ChatGPT. However, adjusting the styling took nearly 2 hours. Initially, I just prompted ChatGPT to "move the new buttons below the original buttons", but even though I gave it lots of the HTML, we kept having trouble communicating. I think it's because the code is too much lines already. It kept giving only snippets of new codes to replace instead of the full script and everytime I said "this didn't work, now it's so-and-so" ChatGPT kept re-writing the code instead of incrementally change things like only certain CSS. Eventually I forced myself to use my own brainpower and experiment with the HTML myself :D. After that I gave clear instructions of what elements to create and where to append the new row of buttons.